Some refactoring and GUI changes
