+++ src/main/java/duke/Data/Parser.java
@@ -30,12 +30,12 @@ public class Parser {
      *
      * @param io
      */
-    public void parseInput(String io, TaskList tasks, Storage storage, ManageStudents students) throws FileNotFoundException, ParseException {
+    public void parseInput(String io, TaskList tasks, Storage storage, ManageStudents students, Schedule schedule) throws FileNotFoundException, ParseException {
         int index = 1;
         String input = io;
         String[] word = io.split(" ");
         String cmd = word[0];
-        Schedule schedule = new Schedule(".\\src\\main\\java\\duke\\Module\\timeslots.txt");
+
         switch (cmd) {
 
             case "list":
