+++ src/main/java/duke/Data/timeslots.txt
@@ -1,2 +1,3 @@
 Swimming-05/10/2019 1500-05/10/2019 1600-pool
 Soccer-05/10/2019 1200-05/10/2019 1300-field
+100-05/10/2019 1500-05/10/2019 1600-pool
