+++ src/main/java/duke/Duke.java
@@ -1,15 +1,20 @@
 package duke;
 
 import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.net.URL;
 import java.text.ParseException;
 import java.util.Scanner;
 
 import duke.Data.Storage;
 import duke.GUI.ViewModules;
+import duke.Module.Schedule;
 import duke.Sports.ManageStudents;
 import duke.Task.*;
 
 import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
@@ -19,6 +24,7 @@ public class Duke extends Application {
     private Storage storage;
     private TaskList tasks;
     private ManageStudents students;
+    private Schedule schedule;
     public static Stage window;
 
     public Duke() {
@@ -26,11 +32,12 @@ public class Duke extends Application {
         //Logger logger = logger.getLogger("main")
     }
 
-    public Duke(String filePath) throws FileNotFoundException {
+    public Duke(String filePath) throws FileNotFoundException, ParseException {
         ui = new Ui();
         storage = new Storage(filePath);
         tasks = new TaskList();
         students = new ManageStudents();
+        schedule = new Schedule(".\\src\\main\\java\\duke\\Data\\timeslots.txt");
     }
 
     /**
@@ -53,7 +60,7 @@ public class Duke extends Application {
                     System.exit(0);
                 }
 
-                ui.readCommand(input, tasks, storage, students);
+                ui.readCommand(input, tasks, storage, students, schedule);
             }
         }
     }
@@ -63,13 +70,24 @@ public class Duke extends Application {
      */
     @Override
     public void start(Stage stage) {
-        window = stage;
-        Scene scene = new Scene(ViewModules.layoutHome(), 1280, 720);
-        stage.setScene(scene);
-        stage.setTitle("Sports Manager");
-        stage.setResizable(false);
-        stage.show();
-        window.setScene(scene);
+        try {
+            URL url = Duke.class.getClassLoader().getResource("view/menu.fxml");
+            System.out.println(url);
+            Parent root = FXMLLoader.load(url);
+            stage.setScene(new Scene(root, 1280,720));
+            stage.setTitle("Sports Manager");
+            stage.show();
+
+//            Scene scene = new Scene(ViewModules.layoutHome(), 1280, 720);
+//            stage.setScene(scene);
+//            stage.setTitle("Sports Manager");
+//            stage.setResizable(false);
+//            stage.show();
+//            window.setScene(scene);
+        }
+        catch (IOException e) {
+            System.err.println("Could not find sample.fxml");
+        }
     }
 
 }
