+++ src/main/java/duke/GUI/Controller.java
@@ -1,0 +1,14 @@
+package duke.GUI;
+
+import javafx.fxml.Initializable;
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class Controller implements Initializable {
+
+    @Override
+    public void initialize(URL location, ResourceBundle resources) {
+        System.out.println("View is now loaded!");
+    }
+
+}
