+++ src/main/java/duke/Ui.java
@@ -1,6 +1,7 @@
 package duke;
 
 import duke.Data.Parser;
+import duke.Module.Schedule;
 import duke.Sports.ManageStudents;
 import duke.Task.TaskList;
 import duke.Data.Storage;
@@ -60,9 +61,9 @@ public class Ui {
      * into the Parser class
      * @param input This is the string input defined by the user
      */
-    public void readCommand(String input, TaskList tasks, Storage storage, ManageStudents students) throws FileNotFoundException, ParseException {
+    public void readCommand(String input, TaskList tasks, Storage storage, ManageStudents students, Schedule schedule) throws FileNotFoundException, ParseException {
         Parser parser = new Parser();
-        parser.parseInput(input, tasks, storage, students);
+        parser.parseInput(input, tasks, storage, students, schedule);
     }
 
 }
