--- src/main/java/duke/exceptions/DukeException.java
@@ -1,12 +1,0 @@
-package duke.exceptions;
-
-import java.lang.Exception;
-
-/**
- * Represents duke.exceptions specific to Duke
- */
-public class DukeException extends Exception {
-    public DukeException (String message) {
-        super(message);
-    }
-}
