+++ src/main/resources/view/sample.fxml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<VBox xmlns="http://javafx.com/javafx"
+      xmlns:fx="http://javafx.com/fxml"
+      fx:controller="duke.GUI.Controller"
+      prefHeight="400.0" prefWidth="600.0">
+    <Label text ="I love bacon"/>
+    <Button text = "Submit"/>
+</VBox>
