+++ src/test/java/DukeExceptionTest.java
@@ -1,4 +1,4 @@
-import duke.DukeException;
+import duke.Exceptions.DukeException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
